renamed trackers to tracker, added remove route
