+++ preordain/tracker/models.py
@@ -1,8 +1,7 @@
 from preordain.models import BaseResponse, RespStrings
-from preordain.trackers.schema import CardInfoModel
+from preordain.tracker.schema import CardInfoModel
 
 
 class SuccessfulRequest(BaseResponse):
     resp = RespStrings.card_info
-    info = {"message": "Successful update!"}
     data: CardInfoModel
