+++ preordain/tracker/router.py
@@ -2,13 +2,22 @@ import logging
 
 log = logging.getLogger()
 from preordain.utils.connections import connect_db, send_response
-from preordain.trackers.schema import CardInfoModel
-from preordain.trackers.models import SuccessfulRequest
+from preordain.tracker.schema import CardInfoModel
+from preordain.tracker.models import SuccessfulRequest
 
 # from preordain.trackers.dependencies import duplicate_card
 from fastapi import APIRouter, Response, status, Depends
 
 router = APIRouter()
+
+validate_query = """
+    SELECT
+        name, card_info.sets.set_full, id, scrape_sales
+    FROM card_info.info
+    JOIN card_info.sets
+        ON card_info.sets.set = card_info.info.set
+    WHERE uri = %s
+"""
 
 
 @router.post(
@@ -27,9 +36,19 @@ async def add_card_to_track(response: Response, card: CardInfoModel):
     """
     cur.execute(query, (card_dict["uri"],))
     conn.commit()
-    conn.close()
-    response.status_code = status.HTTP_201_CREATED
 
+    cur.execute(validate_query, (card_dict["uri"],))
+    data = cur.fetchone()
+
+    if data["scrape_sales"]:
+        conn.close()
+        response.status_code = status.HTTP_201_CREATED
+
+        return SuccessfulRequest(
+            status=response.status_code,
+            info={"Successfully stopped tracking "},
+            data=data,
+        )
     return SuccessfulRequest(status=response.status_code, data=card)
 
 
@@ -49,7 +68,16 @@ async def remove_card_to_track(response: Response, card: CardInfoModel):
     """
     cur.execute(query, (card_dict["uri"],))
     conn.commit()
-    conn.close()
-    response.status_code = status.HTTP_201_CREATED
 
-    return SuccessfulRequest(status=response.status_code, data=card)
+    cur.execute(validate_query, (card_dict["uri"],))
+    data = cur.fetchone()
+
+    if not data["scrape_sales"]:
+        conn.close()
+        response.status_code = status.HTTP_201_CREATED
+
+        return SuccessfulRequest(
+            status=response.status_code,
+            info={"Successfully stopped tracking "},
+            data=data,
+        )
